+++ src/roblox_api/mod.rs
@@ -1,3 +1,4 @@
+mod cframe;
 mod instance;
 
 use std::sync::Arc;
@@ -10,6 +11,7 @@ use crate::{
     value::{Color3Value, Vector3Value, Vector3int16Value},
 };
 
+use cframe::CFrameUserData;
 pub use instance::LuaInstance;
 
 pub struct RobloxApi;
@@ -20,6 +22,7 @@ impl RobloxApi {
         context.globals().set("Vector3", Vector3)?;
         context.globals().set("Vector3int16", Vector3int16)?;
         context.globals().set("Color3", Color3)?;
+        context.globals().set("CFrame", CFrameUserData)?;
 
         Ok(())
     }