+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased Changes
 
 * Added Instance:GetFullName() ([#49](https://github.com/rojo-rbx/remodel/pull/49))
+* Added `Instance:FindFirstChildOfClass()` ([#50](https://github.com/rojo-rbx/remodel/pull/50))
+* Added support for CFrame ([#48](https://github.com/rojo-rbx/remodel/pull/48))
 * Added support for Vector3, and improved Vector3int16 ([#46](https://github.com/rojo-rbx/remodel/pull/46))
 * Added Color3.fromRGB(red, blue, green) ([#44](https://github.com/rojo-rbx/remodel/pull/44))