+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 
+* Added Region3 and Region3int16 types ([#51](https://github.com/rojo-rbx/remodel/pull/51))
 * Added `Instance:FindFirstChildOfClass()` ([#50](https://github.com/rojo-rbx/remodel/pull/50))
 * Added support for CFrame ([#48](https://github.com/rojo-rbx/remodel/pull/48))
 * Added support for Vector3, and improved Vector3int16 ([#46](https://github.com/rojo-rbx/remodel/pull/46))